+++ NEWS.md
@@ -1,5 +1,9 @@
 # uom.dat 0.1.0.9000 (development version)
 
+## New features
+
+* Pattern supports are now computed and added to pattern data frames as another characteristic.
+
 
 
 # uom.dat 0.1.0 (2022-05-06)